requirements.txt
asgiref==3.10.0
contourpy==1.3.3
cycler==0.12.1
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.0
django-environ==0.12.0
django-startproject==1.0a0
djangorestframework==3.16.1
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
narwhals==2.10.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
pyreadstat==1.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
rest-framework-simplejwt==0.0.2
savReaderWriter==3.4.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
librosa
numpy
scipy
soundfile
transformers
torch
sentencepiece
tensorflow
scikit-learn
pandas
matplotlib
seaborn
openai-whisper
ffmpeg-python